Refactor note management UI components and replace Dialog with NoteInputDialog
