+++ src/frontend/src/features/manageNote/ui/EditNote.tsx
@@ -1,9 +1,8 @@
 import { type ReactElement, type FC } from 'react';
-import { useAppDispatch, useAppSelector } from '@/app/store';
+import { useAppDispatch } from '@/app/store';
 import { type NoteItem } from '@/entities/note';
-import { Button, Dialog } from '@/shared/ui';
-import { actions, selectors } from '../model';
-import { NoteInputFlow } from './NoteInputFlow';
+import { actions } from '../model';
+import { NoteInputDialog } from './NoteInputDialog';
 
 interface Props {
   note: NoteItem;
@@ -11,12 +10,6 @@ interface Props {
 }
 
 export const EditNote: FC<Props> = ({ note, renderTrigger }) => {
-  const dialogTitle = useAppSelector(selectors.dialogTitle);
-  const dialogVisible = useAppSelector(state => selectors.editDialogVisible(state, note));
-  const activeFormId = useAppSelector(selectors.activeFormId);
-  const submitText = useAppSelector(selectors.submitText);
-  const submitDisabled = useAppSelector(state => state.manageNote.submitDisabled);
-  const isSubmitting = useAppSelector(state => state.manageNote.isSubmitting);
   const dispatch = useAppDispatch();
 
   const handleDialogOpen = (): void => {
@@ -36,37 +29,10 @@ export const EditNote: FC<Props> = ({ note, renderTrigger }) => {
     );
   };
 
-  const handleDialogClose = (): void => {
-    dispatch(actions.noteDraftDiscarded());
-  };
-
   return (
     <>
       {renderTrigger(handleDialogOpen)}
-      <Dialog
-        pinToTop
-        renderMode="fullScreenOnMobile"
-        title={dialogTitle}
-        opened={dialogVisible}
-        onClose={handleDialogClose}
-        content={<NoteInputFlow date={note.date} />}
-        renderCancel={props => (
-          <Button {...props} type="button" onClick={handleDialogClose}>
-            Cancel
-          </Button>
-        )}
-        renderSubmit={props => (
-          <Button
-            {...props}
-            type="submit"
-            form={activeFormId}
-            disabled={submitDisabled}
-            loading={isSubmitting}
-          >
-            {submitText}
-          </Button>
-        )}
-      />
+      <NoteInputDialog date={note.date} mealType={note.mealType} note={note} />
     </>
   );
 };
